perf(tui): precompute derived entry fields and version-key render caches
Compute per-entry `created` timestamp, `word_count`, and `search_haystack`
once at load so the grouping, stats, row-builder, and fuzzy-search paths
never re-parse dates or re-tokenize bodies on the render/keystroke path.

Group the render memo caches (rows, rendered body, journal stats) into a
`RenderCaches` struct with two version counters: `entries_version` (bumped
when entries change) keys the body/stats caches, while `rows_version`
(bumped on entries or search-hit changes) keys the row cache. A search
recompute now bumps only `rows_version`, so the more expensive body and
stats memos survive query edits.

Widen the scrollbar scroll parameter from u16 to usize for pixel-offset
scrolling.

+++ src/tui/hit_test.rs
@@ -32,14 +32,14 @@ pub(crate) fn entry_index_at(
     geometry: EntryListGeometry,
     x: u16,
     y: u16,
-    scroll: u16,
+    scroll: usize,
     rows: &[EntryRowMeta],
 ) -> Option<usize> {
     if !point_in_rect(geometry.panel.content, x, y) {
         return None;
     }
 
-    let target_y = scroll as usize + y.saturating_sub(geometry.panel.content.y) as usize;
+    let target_y = scroll + y.saturating_sub(geometry.panel.content.y) as usize;
     let mut row_y = 0usize;
     for row in rows {
         let next_y = row_y.saturating_add(row.height as usize);
